refactor: :recycle: refactor book tabs and add new endpoint
