+++ lib/repository/book.dart
@@ -38,6 +38,39 @@ class BookRepository {
     }
   }
 
+   Future<ApiResponse> getBooksPopular() async {
+    ApiResponse response = ApiResponse();
+    List books = [];
+    try {
+      final request = HttpHelper.get('/book/all/most/liked');
+
+      await request.then((result) {
+        response.status = result.data["status"];
+        if (response.status == true) {
+          var objects = result.data["data"];
+
+          for (var book in objects) {
+            var bookObject = Book.fromMap(book);
+            books.add(bookObject);
+          }
+
+          response.data = books;
+        } else {
+          response.error = result.data["error"];
+        }
+      }).catchError((e) {
+        response.status = false;
+        response.error = HttpHelper.getError(e);
+        print(e);
+      });
+
+      return response;
+    } catch (error) {
+      print(error);
+      return response;
+    }
+  }
+
   Future<ApiResponse> getBookById(int id) async {
     ApiResponse response = ApiResponse();
     try {
